refactor: move resolutions mapper
